Solved issue with functions, main.py working now. Need to enhance flexibility and functionalities

+++ STATISTICS/src/bindings.cpp
@@ -1,5 +1,6 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
+#include <pybind11/functional.h>
 #include <pybind11/stl_bind.h>
 #include "DataFrameWrapper.hpp"
 
@@ -8,95 +9,95 @@ namespace py = pybind11;
 PYBIND11_MODULE(DataFrameWrapper, m) {
     m.doc() = "Python bindings for DataFrameWrapper";
 
-    py::class_<Toolbox::dfw::DataFrameWrapper<double>>(m, "DataFrameWrapper")
+    py::class_<Toolbox::dfw::DataFrameWrapper<int>>(m, "DataFrameWrapper")
         .def(py::init<>())
         .def(py::init<const char *, const char *>(),
              py::arg("inputFilename"), py::arg("outputFilename"))
 
-        .def("load_and_read_file", &Toolbox::dfw::DataFrameWrapper<double>::loadAndReadFile,
+        .def("load_and_read_file", &Toolbox::dfw::DataFrameWrapper<int>::loadAndReadFile,
              "Load and read the input file.")
 
-        .def("get_col_index", &Toolbox::dfw::DataFrameWrapper<double>::getColIndex,
+        .def("get_col_index", &Toolbox::dfw::DataFrameWrapper<int>::getColIndex,
              py::arg("columnName"),
              "Get the index of a column by name.")
 
-        /* .def("columns_by_name", &Toolbox::dfw::DataFrameWrapper<double>::template columns<double>, */
+        /* .def("columns_by_name", &Toolbox::dfw::DataFrameWrapper<int>::template columns<int>, */
         /*      py::arg("columnName"), */
         /*      "Retrieve a column by name."); */
 
-        .def("columns_by_name", [](Toolbox::dfw::DataFrameWrapper<double> &self, const char *columnName) {
+        .def("columns_by_name", [](Toolbox::dfw::DataFrameWrapper<int> &self, const char *columnName) {
                 return self.columns<double>(columnName); // Explicit template argument
             },
             /* py::arg("columnName"), */
             "Retrieve a column by name.")
 
 
-        /* .def("columns_by_index", &Toolbox::dfw::DataFrameWrapper<double>::template columns<double>, */
+        /* .def("columns_by_index", &Toolbox::dfw::DataFrameWrapper<int>::template columns<int>, */
         /*      py::arg("col_index"), */
         /*      "Retrieve a column by index.") */
         
-        .def("columns_by_index",[](Toolbox::dfw::DataFrameWrapper<double> &self, size_t col_index) {
+        .def("columns_by_index",[](Toolbox::dfw::DataFrameWrapper<int> &self, size_t col_index) {
                  return self.columns<double>(col_index); // Explicit template argument
             },
 /* py::arg("col_index"), */
            "Retrieve a column by index.")
 
-        /* .def("columns_by_entry", &Toolbox::dfw::DataFrameWrapper<double>::template columns<double>, */
+        /* .def("columns_by_entry", &Toolbox::dfw::DataFrameWrapper<int>::template columns<int>, */
         /*      py::arg("col_index"), py::arg("row_index"), */
         /*      "Retrieve a specific entry in a column.") */
         
           
-        .def("columns_by_entry",[](Toolbox::dfw::DataFrameWrapper<double> &self, size_t col_index, size_t row_index) {
+        .def("columns_by_entry",[](Toolbox::dfw::DataFrameWrapper<int> &self, size_t col_index, size_t row_index) {
                 return self.columns<double>(col_index, row_index); // Explicit template argument
             },
              /* py::arg("col_index"), py::arg("row_index"), */
              "Retrieve a specific entry in a column.")
 
-        /* .def("columns_by_slice", &Toolbox::dfw::DataFrameWrapper<double>::template columns<double>, */
+        /* .def("columns_by_slice", &Toolbox::dfw::DataFrameWrapper<int>::template columns<double>, */
         /*      py::arg("col_index"), py::arg("start"), py::arg("stop"), */
         /*      "Retrieve a slice of a column.") */
 
-        .def("columns_by_slice", [](Toolbox::dfw::DataFrameWrapper<double> &self, size_t col_index, size_t start, size_t stop) {
+        .def("columns_by_slice", [](Toolbox::dfw::DataFrameWrapper<int> &self, size_t col_index, size_t start, size_t stop) {
                 return self.columns<double>(col_index, start, stop); // Explicit template argument
             },
              /* py::arg("col_index"), py::arg("start"), py::arg("stop"), */
              "Retrieve a slice of a column.")
 
-        .def("standard_deviation", &Toolbox::dfw::DataFrameWrapper<double>::template StandardDeviation<double>,
+        .def("standard_deviation", &Toolbox::dfw::DataFrameWrapper<int>::template StandardDeviation<double>,
              py::arg("columnName"),
              "Calculate the standard deviation of a column.")
 
-        .def("mean", &Toolbox::dfw::DataFrameWrapper<double>::template Mean<double>,
+        .def("mean", &Toolbox::dfw::DataFrameWrapper<int>::template Mean<double>,
              py::arg("columnName"),
              "Calculate the mean of a column.")
 
-        .def("variance", &Toolbox::dfw::DataFrameWrapper<double>::template Variance<double>,
+        .def("variance", &Toolbox::dfw::DataFrameWrapper<int>::template Variance<double>,
              py::arg("columnName"),
              "Calculate the variance of a column.")
 
-        .def("median", &Toolbox::dfw::DataFrameWrapper<double>::template Median<double>,
+        .def("median", &Toolbox::dfw::DataFrameWrapper<int>::template Median<double>,
              py::arg("columnName"),
              "Calculate the median of a column.")
 
-        .def("correlation", &Toolbox::dfw::DataFrameWrapper<double>::template Correlation<double>,
+        .def("correlation", &Toolbox::dfw::DataFrameWrapper<int>::template Correlation<double>,
              py::arg("columnName1"), py::arg("columnName2"),
              "Calculate the correlation between two columns.")
 
-        .def("frequency_count", &Toolbox::dfw::DataFrameWrapper<double>::template frequencyCount<double>,
+        .def("frequency_count", &Toolbox::dfw::DataFrameWrapper<int>::template frequencyCount<double>,
              py::arg("columnName"),
              "Get the frequency count of a column.")
 
-        .def("classify", &Toolbox::dfw::DataFrameWrapper<double>::template Classify<double>,
+        .def("classify", &Toolbox::dfw::DataFrameWrapper<int>::template Classify<double>,
              py::arg("columnName1"), py::arg("categories"), py::arg("conditions"),
              "Classify column data based on conditions.")
 
 #ifdef INTERPOLATION_MODULE
-        .def("interpolate", &Toolbox::dfw::DataFrameWrapper<double>::template Interpolate<double>,
+        .def("interpolate", &Toolbox::dfw::DataFrameWrapper<int>::template Interpolate<double>,
              py::arg("columnName1"), py::arg("columnName2"), py::arg("categories"), py::arg("conditions"),
              "Interpolate column data based on conditions.")
 #endif
 
-        .def("get_info", &Toolbox::dfw::DataFrameWrapper<double>::getInfo,
+        .def("get_info", &Toolbox::dfw::DataFrameWrapper<int>::getInfo,
              "Retrieve basic information about the DataFrame.");
 
     py::bind_map<std::unordered_map<std::string, unsigned long>>(m, "UnorderedMap");
